examples/rag-app/requirements.txt
# RAG Application Dependencies
# Python 3.9+

# Core LLM Framework
langchain==0.1.9
langchain-community==0.0.20
langchain-anthropic==0.1.4
langchain-openai==0.0.6

# Vector Database
pinecone-client==3.0.1
pinecone-text==0.1.5

# Document Processing
pypdf==4.0.1
python-docx==0.8.11
markdown==3.5.1

# LLM Providers
anthropic==0.7.1
openai==1.3.5

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
tqdm==4.66.1

# Optional: Performance & Monitoring
langsmith==0.0.68
